third_party/WebKit/LayoutTests/platform/win/external/wpt/hr-time/window-worker-time-origin-expected.txt

This is a testharness.js-based test.
FAIL Worker time origin is approximately its creation time assert_less_than: worker performance.now() must be very close to its creation time expected a number less than 500 but got 1354.9850000000001
Harness: the test ran to completion.
